requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
SpeechRecognition==3.8.1
groq==0.0.1
requests==2.27.0
gunicorn==20.1.0